requirements.txt
django==1.8
django-angular==0.7.3
django-any-imagefield==0.8.1
django_compressor==1.3
django-nose==1.4
django-tastypie==0.11.0
ipdb==0.8
ipython==3.1.0
Pillow==2.4.0
